services/api/requirements.txt
fastapi==0.68.2 # Fix for https://github.com/tiangolo/fastapi/issues/4025
feedparser==6.0.10
firebase_admin==6.2.0
google-cloud-logging==2.7.0
google-cloud-pubsub==2.9.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
ratelimiter==1.2.0.post0
requests==2.26.0
starlette-context==0.3.3
uvicorn==0.15.0
